add structured configuration
we want to make the target formats configurable. as we want to keep
the code agnostic to the differently named target formats structured
configuration is a way to achieve this. in the target datastructure
the formats are kept in a map. I chose toml as the format since json
is to verbose, properties are not expressive enough and yaml just sucks.